add cloze-respecting syntax highlighting
Use the highlight extension from pymdownx to highlight code blocks
that are identified by the superfences extension.

Implement a custom lexer for Pygments that first strips out the cloze
markers from the source code since these have no meaning in the language
the code snippet is written in, so they usually make the syntax of the
code snippet invalid.

Monkey patch the pygments library so that when a lexer is requested
for a language, a delegating lexer is returned instead that first
strips out the clozes, and then delegates lexing the remaining code
to the lexer that would originally have been returned by
get_lexer_by_name.